add delete staff member feature

+++ src/main/java/com/library/LibMgmt2025/controller/StaffController.java
@@ -2,6 +2,7 @@ package com.library.LibMgmt2025.controller;
 
 import com.library.LibMgmt2025.dto.BookDto;
 import com.library.LibMgmt2025.dto.StaffDto;
+import com.library.LibMgmt2025.exception.StaffNotFoundException;
 import com.library.LibMgmt2025.service.StaffService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -28,8 +29,20 @@ public class StaffController {
     }
     @DeleteMapping
     public ResponseEntity<Void> deleteStaff(@RequestParam("staffId") String staffId){
-        staffService.deleteStaff(staffId);
-        return ResponseEntity.noContent().build();
+        if(staffId == null){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        try {
+            staffService.deleteStaff(staffId);
+            return ResponseEntity.noContent().build();
+        }catch (StaffNotFoundException e){
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }catch (Exception e){
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+
     }
 
     @PatchMapping(value = "/{staffId}", consumes = MediaType.APPLICATION_JSON_VALUE)
